Revert "dma-coherent: add dma exact size alloc/release functions"
Remove downstream heap management code from core kernel to move closer
to upstream.
This reverts commit b6a2bea0d5f910962e3856ab6412602fbef43f33.

Bug 3415117

Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2626580
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit

+++ include/linux/dma-map-ops.h
@@ -173,7 +173,7 @@ int dma_mmap_from_dev_coherent(struct device *dev, struct vm_area_struct *vma,
 
 void *dma_alloc_from_global_coherent(struct device *dev, ssize_t size,
 		dma_addr_t *dma_handle);
-int dma_release_from_global_coherent(size_t order, void *vaddr);
+int dma_release_from_global_coherent(int order, void *vaddr);
 int dma_mmap_from_global_coherent(struct vm_area_struct *vma, void *cpu_addr,
 		size_t size, int *ret);
